capsule_uuid: return hex string, not UUID obj
In Python 2, uuid.hex is bytes; in Python 3, unicode. It's a dynamic
descriptor, so it can't easily be overridden without getting clever. I thought
about it and decided I don't really like objects that wrap primitives anyhow.
Better to have capsule_uuid() return a unicode string containing the
hexadecimal representation of the UUID.

Change-Id: Ib91aa58449d685d9f31b42130f244edc631948e1