(Bug 46011) Log $_SERVER['HTTP_HOST'] as webHost
I'm not sure that this is a fully satisfactory solution, but it helps
disambiguate the particular case cited in the bug (beta cluster vs.
production), and it's a change that is very nearly free to make.

The capsule schema revision changes the description of 'webHost' to
indicate that it is also logged server-side, when available. The diff is
viewable here:

<http://meta.wikimedia.org/w/index.php
?title=Schema%3AEventCapsule&diff=5315751&oldid=5262883>

Change-Id: I355b0f662083cea9913617161719bf43494fd615