Default MariaDB character encoding for EL data: utf8 -> utf8mb4
This change sets the default character encoding for MySQL / MariaDB
EventLogging data to 'utf8mb4' (was: 'utf8'), adding support for characters
above the Base Multilingual Plane. Deployment will require manual migration of
existing data in the database.

One of the consequences of this migration is that the previous default size for
string columns is not longer appropriate, since the columns it generates are
not indexable by InnoDB, which will not index columns beyond 767 bytes. This
change therefore amends the default size to be 191, which is the maximum size a
utf8mb4 string column can be and still remain indexable.

Finally, as a way of not being blocked on deployment of I8fdcc046d, this change
adds a live hack that substitutes 'utf8mb4' for 'utf8' in database connection
strings. The hack can be removed once I8fdcc046d is deployed.

FIXME: Database setup instructions and minimum requirements should be
documented.

Change-Id: Ia94f2c2155de5fb9031a8164306720e06455cced