Add coroutine-based, URI-configured generic event processor
Rather than have several scripts (each with different command-line argument
semantics) that each do the same basic thing -- consume an event stream into
some data store -- it is better to have a script that takes an input source and
an output source and correctly delegates to the right handler for each.

Making handler registration dynamic (using the @writes and @reads decorators)
makes it possible to have a plug-in architecture for event writers. A future
change-set will load Python files from some designated path, allowing people to
simply write a coroutine to write out to some data store and register it with
EventLogging, which will have the effect of making it available to the main
'eventlogger' script.

'eventlogger' deprecates json2mongo, json2sql and log2file. Because we are the
only users of this module in production, I am not publicizing the deprecation.
I'll just remove the old scripts once I've tested this in production.

Change-Id: I301565fe87e24a00b6ae9b2238f0f0dca36c608d