Add 'urisplit' to gloss over urlparse inconsistencies
This patch adds a 'urisplit' method which wraps over the inconsistencies in the
handling of query strings and fragments in URIs that use custom schemes.
Python's historic behavior was to not split off a query string or fragment from
the path. This was changed in Python 2.7.4 and was backported to certain builds
of 2.7.3 (Debian) but not all (Ubuntu, Fedora).

Change-Id: I19874cb90c974a2fa29531056bb75304bddf2b87