Use HMAC SHA1 with rotating key to hash IP addresses
The current behavior is to generate a 'salt' (really a key) at runtime and to
use it continuously for the lifetime of the program. The lifespan of the key
and the cheapness of the hash function make it easier to attack. This change
makes EventLogging scramble IPs by generating an HMAC SHA1 with the IP address
as the message and a random byte string as the key. The key rotates every 90
days.

Change-Id: Ibbaf182c2e165911b51755da51fbf1769cf67edf