SQL consumer: make `uuid` column unique & not null
EventLogging guarantees that each event is tagged with a unique UUID, but for
performance reasons the guarantee was not hitherto expressed as a constraint in
the database. The absence of a constraint makes it substantially more difficult
to backfill data from the file-based logs in the case of a database outage,
because it means we need to explicitly check for the existence of each UUID
that may be missing from the database if we want to avoid dupes.

Sean thinks the performance impact of having a unique constraint will be
negligible at the rate of events that EventLogging is handling.

Change-Id: Ifb2b8ff96a549a3c62d0ad75b156d088d1e13e0e