Don't print out full exception details on validation error
Currently, a jsonschema validation error will cause the entire schema to be printed out,
as well as any other information about the validation error.  We don't need to see the
schema for every bad event in the error logs, so only print the raw event and the
validation exception error message.

Bug: T93185
Change-Id: I855c39d01c2b4be727c46ee1e168abf4949c55c7