Processor's output-invalid now sends an error json object with the error message and the raw event
I also changed the default kafka writer topic and key to not include any interpolatable keys.
This way the defaults will work with any json objects, and won't throw a KeyError if
keys like 'schema' and 'revision' don't exist in the event.

This means that kafka writers that want to produce to schema based topics will have to
always manually specify topic and key format strings, but they probably would
do that anyway.

Change-Id: I792463d29fc222353a662bef3792faec0b3ddd8e