eventlogging-processor can now use the same output URI as the output invalid writer
This allows you to configure the processor to write to Kafka schema based topics
with a single output URI for both valid and invalid events, as long as you want
them all written to a similar naming scheme.  Since invalid events are written
using the EventError schema, they can be written to a topic name that is
interpolated using the same format string as valid topics, e.g.

  kafka:///broker.example.net?topic=eventlogging_%(schema)s&key=%(schema)s_%(revision)s

This commit also reverts the previous commit's change to set the default kafka writer
topic and key to use event capsule string interpolation, as it is expected that you
will be writing event objects to both the valid and invalid outputs URIs.  Now, if
you want the the output URI of both valid and invalid events to be the same as
the example above, you may omit the topic and key parameters.  I.e.

  kafka:///broker.example.net

is sufficent.

Change-Id: I4f15caabe5e4f93bb79442086800b3493147a50f