Write different EventError code for validation errors
Now that errors are encapsulated in the EventError schema, we can
differentiate between validation and other types of processor errors.
This will help eventlogging maintainers tell if an event is not
being produced propertly because of a validation error, or because
of something more annoying, like data truncation.

This will require the a schema change to EventError to add
'validation' as a possible enum value for error code.

Change-Id: Ia55906db7b6214e6d4b6c20477deb2979524cbe5