Add the Swagger spec for the proxy service
This also adds the ability for automated swagger spec testing
via the x-ample request spec.  In order to do this
for all eventlogging-services, indepdendent of the
user provided --schemas-path and --topic-config,
all eventlogging-services must have the
test.event topic config and test_event schema
configured.  This patch hardcodes this topic config
and schema and always adds them to an eventlogging-service's
topic_config and schema_cache.

Change-Id: I9a124e8aa6fcdd7491e652705e9e54a19e85909c