Add a format specifier that indicates to parser to ignore the value
This change introduces a format specifier %o - that will cause it's value
to be omitted/ignored by the eventlogging parser. This is a step towards
dropping ClientIPs from EventCapsule. We will switch the client side processor
format string to use %o instead of %h on puppet to start dropping IPs at the
Eventlogging end. We can safely merge/deploy this without causing any change
to the current system.

Bug: T128407
Change-Id: Ib1af6cb2270502abec9cce4ec72240e89064a261