Upgrade kafka writer to kafka-python 1.2.5+, add new kafka-python:// reader to test
https://github.com/dpkp/kafka-python/pull/783 fixes a performance degradation
specifically related to synchronous producing, so this comit will
work best with kafka-python 1.2.6 or greater.

This also DRYs the logic for getting kafka client and group ids,
and adds an Event.serialize method to DRY up converting
events to JSON utf-8 strings.

Note that consumer balancing with kafka-python requires Kafka 0.9+

Bug: T133779
Change-Id: I3df00e9219f0a80c0805edd1df01c12fe8697341