Properly reload tornado child processes during SIGHUP
If num_processes != 1, tornado will fork children to process HTTP requests.
The SIGHUP handler that was registered on the main process had no effect on
these child processes.  With this change, a SIGHUP on the parent process
will cause a SIGHUP to also be sent to each child process.
In addition, the actual reloading of schemas and topic configs is now
done as part of each processes' tornado ioloop, rather than directly
from the signal handler.

This also adds pid information to the default logging format.

Bug: T140870

Change-Id: I73289e1a413cc38445db5b52d018d41e1a0da418