ID schemas with path context, add /v1/schemas endpoint to eventlogging-service
/v1/schemas makes it easier to see what schemas eventlogging-service has loaded.

This is a change in the way schemas are loaded and named and IDed in eventlogging.
Previously, schema names were looked up using only the parent directory
of the schema.  That is mediawiki/page_create/1.yaml would end up being
named 'page_create'.  Now, the name will be 'mediawiki/page_create'.

This allows nesting of schemas without losing context in schema names or
schema_uris.  Note that this is a breaking change.  eventlogging-service
topic-config should be updated to reflect the change in schema name.

Bug: T134502
Change-Id: I3b68ab173b7a16e4f5fbb8abfc48b8a6f6ec037c