Change UA string to JSON map
Use ua_parser to generate a JSON object with properties obtained from
the user agent string. This change increases the length of user agent field
in order to deploy it such user agents are not truncated we need
to increase length of column. Please see:
https://phabricator.wikimedia.org/T160454

Custom code parses the WMF app version

Bug: T153207
Change-Id: I4480619233bcc530754e08b4b9081b45d9d2e785