Build SQL table name from topic if set, else use schema name
In order to support schema sharing in eventbus topics, we need to
be able to insert events from different topics into different tables,
even if events in those topics are of the same schemas.
This will likely be the case for a new page create event,
as it might re-use the revision create schema.

Bug: T150369
Change-Id: I16687fc4e006ce961426c17343511d90019aac0d