Fix mysql handler scid grouping
The mysql handler groups events by scid before it sends them to
store_sql_events. store_sql_events assumes all events in the batch
belong to the same schema, and gets the table it should insert
them all from the first event in the batch.

This is a problem with events that come from eventBus, because
they can share a scid. If they do, the mysql handler will group
them together in the same batch and send them together to
store_sql_events, who will determine the table to insert into
by the topic of the first event in the batch. This causes that
the whole batch of mixed events gets inserted into the same table.

This patch fixes the grouping of the events by adding the topic
to the group key (batch_key).

Bug: T150369
Change-Id: Ie714f4fcf7bb66614e8f4f89dbf7a66e6464aaba