Support revisionless schema events to eventlogging service
Currently, the MediaWiki EventBus extenstion does not set schema_uri
on any events it sends.  Instead, it sets topic, and lets
eventlogging-service set schema_uri appropriately with the latest
schema,revision configured for the intended topic.

This change makes eventlogging service support events that set
the schema name (either in the schema_uri, or top level 'schema')
without a revision.  This isn't something that EventBus
extension needs to do, but we should allow producers to
set the schema they intend the event to be validated with, if they
so choose.  This will still fail if they try to set the schema
to one not configured for the intended topic.

Change-Id: Ica850d66eb1dc659f67af604cd24cfd1d87b1ddc