Produce kafka-python messages with timestamp set to Event time
This adds an Event.timestamp method to convert the extracted
event datetime to a utc unix timestamp.  It also
adds the assumption that any timezoneless datetimes parsed
out of events are in UTC.  UTC Event datetimes that have their
dt's iso 8601 string formatted will have the timezone offset
encoded as 'Z' instead of '+00:00'.

Bug: T196407
Change-Id: I826cd2c34c5b34195018d4f0b2c2247520f847a8