Set logging level to all kafka libraries to ease debugging sessions
During the last Eventlogging outage a lot of time was spent to figure
out how to properly enable debug logging for all the Kafka Python
libs (excluding confluent-kafka-python that uses its own settings).

This change also adds class name in which the logging happens and
its level, to easily spot what library is responsible for the log.

Bug: T200765
Change-Id: Ie8680c47fd0a070c214e285f5c5fef1fd5863c52