event.py: avoid encoding a json string
In Python 3 a string is already encoded with utf-8 by default.
The encode() function creates a bytes string, that shouldn't
be used for "application" processing but only when doing stuff
like sending bytes over the network or similar.

In the current case, json.dumps(a) fails if a dictionary contains
a key value pair with a byte encoded string. This does not happen
in Python 2.

Bug: T233231
Change-Id: Ieed6404b2a6c073b8adc05512310b74eedb433a0